home and shop shit

+++ OOPP/app/src/main/java/com/example/sauronsarmy/oopp/StatsPresenter.java
@@ -13,7 +13,7 @@ public class StatsPresenter implements StatsPresenterInterface {
     }
 
     @Override
-    public int getMoneyPerSecond() {
+    public double getMoneyPerSecond() {
         return playerModel.getMoneyPerSecond();
     }
