+++ OOPP/app/src/main/java/com/example/sauronsarmy/oopp/StatsPresenterInterface.java
@@ -12,5 +12,5 @@ interface StatsPresenterInterface {
 
     int getMoneyAmount();
 
-    int getMoneyPerSecond();
+    double getMoneyPerSecond();
 }
